fix(cdk/testing): require at least one argument for locator functions (#23619)
Currently locator functions accept a spread argument which technically allows for zero selectors to be passed in. This can result in weird runtime errors.

These changes add a runtime error if no selectors are passed in.

Note that a previous iteration of these changes tried to enforce this with typings, but it resulted in breaking changes.