refactor: use explicit specifier for form field control harness re-export
Instead of using a star export, we should use an explicit specifier
for the `@angular/material/form-field/testing/control` re-export.

This satisfies Microsoft's API extractor which will complain as
it cannot resolve the entry-point. Within an API golden test, first-party
packages/entry-points are always considered external and therefore not resolvable.