build: avoid duplication of mdc dependencies by generating `package.json` (#24697)
As part of the Yarn PNP changes we moved from `peerDependencies` in the
experimental package to actual `dependencies` on all individual MDC
packages we rely on. This is necessary for Yarn PnP and conceptually
also makes the package more correct/safe.

This resulted in duplication of the MDC dependencies, making MDC
updates more complicated. We can avoid duplication by auto-inserting
the dependencies. That way we keep two places for the MDC deps:

* The project `package.json`
* `packages.bzl` for the Bazel analysis phase/Starlark.