fix(multiple): enable easy extension of harnesses (#24878)
Add the capability to extend a harness without having to duplicate the `with` method
used to load it. This was necessary before because the returned predicate was of type
the parent harness which resultet in loading a harness instance of the parent type
instead of the derived class unless the method is overridden.