build: invalidate github action cache to prune old files (#24896)
With Angular v14, and the CLDR update one of the files got renamed
just in casing. Yarn 1.x. does not seem to prune the old files
causing two files to have the same name (ignoring casing). This throws
off Bazel and we need to prune the cache like we did for CircleCI.