fix(google-maps): error when minified through closure (#24897)
In #24865 the access of the global `MarkerClusterer` class changed to go through the `window`. This causes a runtime error for apps minified through Closure, because the name is no longer minified and it doesn't align with the minified class name on the `window`. These changes move around our code so that we reference the clusterer without going through the `window`.