test(cdk/testing): use offsetX and offsetY to verify element coordinates
Currently a couple of the tests use `clientX` and `clientY` plus the element's `ClientRect` to verify click coordinates relative to an element. This can cause flakes in tests, because the `ClientRect` changes depending on the page's scroll position.

These changes use `offsetX` and `offsetY` instead which are actually relative to the element.