build: avoid use of github team for dev-infra codeowner (#25337)
We should just use the explicit names since there aren't a lot
of instances, and this is not changing very often.

The change back then did not bring a lot of benefits but rather made
reviews more annoying since unrelevant PRs had to be unassigned twice
to disappear in the `Requested reviews` view. This is because I'm
sometimes requested explicitly through other ownerships, but then also
through the dev-infra group.