feat(material/form-field): Ability to support custom error message components inside a form field (#25399)
* feat(material/form-field): Ability to support custom error message components inside a form field

Adds an ability to use custom error components tagged with [matError] attributes inside a form field. This is useful when a reusable custom error message template is needed for form fields.

Example:

<mat-form-field>
  <mat-label>Label</mat-label>
  <input matInput>
  <custom-error matError>Error message</custom-error>
</mat-form-field>

* feat(material/form-field): Ability to support custom error message components inside a form field

Adds an ability to use custom error components tagged with [matError] attributes inside a form field. This is useful when a reusable custom error message template is needed for form fields.

Example:

<mat-form-field>
  <mat-label>Label</mat-label>
  <input matInput>
  <custom-error matError>Error message</custom-error>
</mat-form-field>

* feat(material/form-field): Ability to support custom error message components inside a form field

Adds an ability to use custom error components tagged with [matError] attributes inside a form field. This is useful when a reusable custom error message template is needed for form fields.

Example:

<mat-form-field>
  <mat-label>Label</mat-label>
  <input matInput>
  <custom-error matError>Error message</custom-error>
</mat-form-field>

Co-authored-by: kamarouski <kamarouski@google.com>