fix(material/core): default font family not picked up in define-typography-config (#25789)
The `define-typography-config` function was set up so that if a level doesn't have a `font-family`, it would pick up the one that was passed into the function. The problem is that after we switched the typography over to MDC, we started using the `typography-config-level-from-mdc` function which provides a default `font-family` which meant that the default passed into `define-typography-config` was being ignored.

These changes add some logic that allows us to take the default font family, if it is specified.

Fixes #25780.