build: setup action for identifying changes needing to be tested in google (#25931)
* build: update angular shared dev-infra code

* build: update ng-dev configuration to account for deleted config options

* build: setup action for identifying changes needing to be tested in google

This will be the public part of the Angular Presubmit Status service. We
will be able to remove the `merge safe` label- but for now we keep it
until we decide what to do with it. I could also imagine this action
automatically applying such a label for better filtering.

At the same time though, for a caretaker it should likely not matter
because the GitHub status checks are sufficient indicator (now with the
new service) whether a PR is ready or not.

Co-authored-by: Angular Robot <angular-robot@google.com>