fix(material/datepicker): resolve typing issue when extending datepicker input in legacy app (#25988)
In the switch to MDC we kept the reference to `MatFormField` in the datepicker input's constructor which now has a slightly different public API, causing a compilation error for apps that were passing in the legacy form field.

These changes replace the reference with an interface that contains only the common members that the datepicker cares about.

Fixes #25977.